requirements.txt
beautifulsoup4
requests
pandas
pytest
pytest-cov
sqlalchemy
google-auth
gspread
pytz